changes to flow of payments for different fees shown

+++ models/Charge.js
@@ -7,7 +7,7 @@ module.exports = (sequelize, DataTypes) => {
 		{
 			id: { type: DataTypes.UUID, primaryKey: true, allowNull: false },
 			description: DataTypes.STRING,
-			amount: DataTypes.INTEGER,
+			amount: DataTypes.DOUBLE,
 
 			payment_status: {
 				type: DataTypes.STRING,
